requirements.txt
streamlit
openai
requests
httpx
python-dotenv
pydantic
pandas
markdown-it-py
pdfplumber
pymupdf
python-docx

